+++ lib/braintree/error_codes.rb
@@ -385,6 +385,7 @@ module Braintree
       SubscriptionStatusMustBePastDue = "91531"
       TaxAmountCannotBeNegative = "81534"
       TaxAmountFormatIsInvalid = "81535"
+      TaxAmountIsRequiredForAibSwedish = "815224"
       TaxAmountIsTooLarge = "81536"
       ThreeDSecureAuthenticationFailed = "81571"
       ThreeDSecureAuthenticationIdDoesntMatchNonceThreeDSecureAuthentication = "915198"