+++ lib/braintree/payment_method_nonce.rb
@@ -15,6 +15,7 @@ module Braintree
     end
 
     attr_reader :bin_data
+    attr_reader :default
     attr_reader :details
     attr_reader :nonce
     attr_reader :three_d_secure_info
@@ -23,12 +24,14 @@ module Braintree
 
     def initialize(gateway, attributes) # :nodoc:
       @gateway = gateway
-      @nonce = attributes.fetch(:nonce)
-      @type = attributes.fetch(:type)
+      set_instance_variables_from_hash(attributes)
       @details = PaymentMethodNonceDetails.new(attributes[:details]) if attributes[:details]
-      @authentication_insight = attributes.fetch(:authentication_insight, nil)
       @three_d_secure_info = ThreeDSecureInfo.new(attributes[:three_d_secure_info]) if attributes[:three_d_secure_info]
       @bin_data = BinData.new(attributes[:bin_data]) if attributes[:bin_data]
+    end
+
+    def default?
+      @default
     end
 
     def to_s # :nodoc: