+++ spec/spec_helper.rb
@@ -42,6 +42,7 @@ unless defined?(SPEC_HELPER_LOADED)
     HiperBRLMerchantAccountId = "hiper_brl"
     CardProcessorBRLMerchantAccountId = "card_processor_brl"
     FakeFirstDataMerchantAccountId = "fake_first_data_merchant_account"
+    AibSwedenMaMerchantAccountId = "aib_swe_ma"
 
     TrialPlan = {
       :description => "Plan for integration tests -- with trial",